Wrap the reader used in the ScifioImgSource with a UnclosableReaderFilter to prevent him from being closed.
The reason: if an image is opened via the ImgOpener the metadata is reset afterwards (even if just one series of the current image file is read). To read another series of the same file the file has to be parsed again, what might be a bit expensive for some file formats.
Now, as the reader is unclosable, it is closed not before all desired series have been read.

Conflicts:

	org.knime.knip.io/src/org/knime/knip/io/ScifioImgSource.java