WaehlbyCellClumpSplitter: Fix "border-pixel-removal" bug
The watershed results in the border pixels marked as sheds which will be
used for splitting later. This causes the border labels always to be
removed. By extending the image by one and later removing the border of the
watershed result, we can avoid this without compromising on result quality.

Signed-off-by: Squareys <Squareys@googlemail.com>